requirements.txt
requests==2.31.0
google-generativeai==0.8.3
lxml==5.1.0
